requirements.txt
bleak==0.21.1
ruuvitag_sensor==2.3.1
Flask==3.0.3
